+++ src/main/kotlin/com/example/hhplus_ecommerce/domain/payment/PaymentEventListener.kt
@@ -1,5 +1,6 @@
-package com.example.hhplus_ecommerce.event
+package com.example.hhplus_ecommerce.domain.payment
 
+import com.example.hhplus_ecommerce.event.DataPlatformEvent
 import com.example.hhplus_ecommerce.infrastructure.message.producer.MessageProducer
 import com.example.hhplus_ecommerce.infrastructure.message.producer.dto.PaymentDataMessage
 import org.springframework.stereotype.Component
@@ -7,7 +8,7 @@ import org.springframework.transaction.event.TransactionPhase
 import org.springframework.transaction.event.TransactionalEventListener
 
 @Component
-class DataPlatformEventListener(
+class PaymentEventListener(
     private val messageProducer: MessageProducer
 ) {
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
